Multiple Fixes
Bugs did not allow "I believe," "I think," "I feel," etc. to be highlighted. The solution was only highlighting the word "I".
More cleaning up the essays' strings was added. These include the removal of "\"(put in after every single quotation), and the single quotations at the beginning and end of the essays' strings. This was already done at the end when displaying the result, however in doing so, detecting certain words(especially ones at the beginning of a paragraph) was impossible.