Fixed detection bug
Bug did not allow for correct display of the amount of dead words found. The exact reason was that CDead was adding dead words no matter if they already existed, hence causing the bug. The fix restricted CDead to only add dead words to be displayed if they weren't already detected.